refactor(frontend): extract shared SA and WI panel components
- Create ServiceAccountPanel.vue and WorkloadIdentityPanel.vue shared components
- Refactor workspace and project dashboard views to use shared components
- Remove unnecessary hasListPermission check (router already handles permission)
